Fixes based on CR.
